+++ src/contracts/index.ts
@@ -68,3 +68,4 @@ export * from "./EventContractCollectionRunnerDurableFixtureRehearsal";
 export * from "./EventContractCollectionRunnerDurableFixtureRehearsalCoordinator";
 export * from "./EventContractCollectionRunnerDurableFixtureRehearsalEvidence";
 export * from "./EventContractCollectionRunnerRuntimeOwnershipRecovery";
+export * from "./EventContractCollectionRunnerRehearsalOperation";